# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CompanyOS Core API",
    "description": "Phase 1.3 \u2014 dashboard BFF + workspace + auth + hello.",
    "license": {
      "name": "Proprietary",
      "identifier": "Proprietary"
    },
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/finance/expenses": {
      "get": {
        "tags": [
          "finance-expenses"
        ],
        "summary": "GET /api/v1/finance/expenses",
        "operationId": "list_expenses",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "finance-expenses"
        ],
        "summary": "POST /api/v1/finance/expenses",
        "operationId": "submit_expense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitExpenseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/finance/expenses/{id}/decide": {
      "post": {
        "tags": [
          "finance-expenses"
        ],
        "summary": "POST /api/v1/finance/expenses/{id}/decide",
        "operationId": "decide_expense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideExpenseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecideExpenseRequest": {
        "type": "object",
        "required": [
          "approve"
        ],
        "properties": {
          "approve": {
            "type": "boolean"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ExpenseDto": {
        "type": "object",
        "required": [
          "id",
          "status",
          "currency",
          "amount_minor",
          "description",
          "incurred_at",
          "created_at"
        ],
        "properties": {
          "amount_minor": {
            "type": "integer",
            "format": "int64"
          },
          "category_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "incurred_at": {
            "type": "string"
          },
          "receipt_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
      "ExpenseListResponse": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "SubmitExpenseRequest": {
        "type": "object",
        "required": [
          "currency",
          "amount_minor",
          "description"
        ],
        "properties": {
          "amount_minor": {
            "type": "integer",
            "format": "int64"
          },
          "category_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "currency": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "incurred_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "receipt_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}